requirements.txt
pympi-ling
requests
numpy
PyQt5
pydub
https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pygame
python-box
wave
python-resize-image
imageio